+++ e2e/cases/solid/hmr/src/index.jsx
@@ -1,4 +1,4 @@
 import { render } from 'solid-js/web';
-import App from './App';
+import App from './A';
 
 render(() => <App />, document.getElementById('root'));