+++ packages/plugin-solid/src/index.ts
@@ -14,9 +14,10 @@ export function pluginSolid(
     pre: ['rsbuild:babel', 'rsbuild-webpack:babel'],
 
     setup(api) {
-      api.modifyBundlerChain(async (chain, { CHAIN_ID }) => {
+      api.modifyBundlerChain(async (chain, { CHAIN_ID, isProd }) => {
         [CHAIN_ID.RULE.JS, CHAIN_ID.RULE.JS_DATA_URI].forEach((ruleId) => {
           if (chain.module.rules.has(ruleId)) {
+            const rsbuildConfig = api.getNormalizedConfig();
             const rule = chain.module.rule(ruleId);
 
             if (rule.uses.has(CHAIN_ID.USE.BABEL)) {
@@ -27,6 +28,20 @@ export function pluginSolid(
                   require.resolve('babel-preset-solid'),
                   options.solidPresetOptions || {},
                 ]);
+
+                if (!isProd && rsbuildConfig.dev.hmr) {
+                  babelConfig.plugins ??= [];
+                  babelConfig.plugins.push([
+                    require.resolve('solid-refresh/babel'),
+                  ]);
+
+                  chain.resolve.alias.merge({
+                    'solid-refresh': require.resolve(
+                      'solid-refresh/dist/solid-refresh.mjs',
+                    ),
+                  });
+                }
+
                 return babelConfig;
               });
             }