+++ packages/plugin-vue-jsx/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @rsbuild/plugin-vue-jsx
 
+## 0.1.4
+
 ## 0.1.3
 
 ## 0.1.2